+++ config.py
@@ -6,6 +6,7 @@ class Config(object):
         TESTING = False
         SECRET_KEY = os.environ.get("SECRET_KEY")
         SQLALCHEMY_TRACK_MODIFICATIONS = False
+        SQLALCHEMY_DATABASE_URI = os.environ.get("DATABASE_PG_URI")
 
     class Production(Default):
         ENV = "production"